Move configuration, use factory
