+++ src/Configuration/Configuration.cs
@@ -1,8 +1,8 @@
-namespace OpcPlc;
+namespace OpcPlc.Configuration;
 
 using System;
 
-public class Configuration
+public class OpcPlcConfiguration
 {
     /// <summary>
     /// Name of the application.
@@ -66,5 +66,5 @@ public class Configuration
 
     public TimeSpan LogFileFlushTimeSpanSec { get; set; } = TimeSpan.FromSeconds(30);
 
-    public OpcApplicationConfiguration OpcUa { get; set; }
+    public OpcApplicationConfiguration OpcUa { get; set; } = new OpcApplicationConfiguration();
 }
